feat(vite):Implement basic configuration for MPA with environment variables.
