+++ demo.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>PERSENAUT</title>
-    <link rel="stylesheet" href="./assets/css/styles.css">
+    <link rel="stylesheet" href="./src/assets/css/styles.css">
 </head>
 
 <body>
@@ -51,7 +51,8 @@
         </div>
     </div>
 
-    <script src="./assets/js/index.js"></script>
+    <!-- <script src="./src/assets/js/index.js"></script> -->
+    <script type="module" src="./src/assets/js/index.js"></script>
 
 </body>
